+++ doc/scripting/variables.qml
@@ -12,7 +12,8 @@ Script {
     property string myText;
     property int myInt;
     property string myFile;
-    
+    property string mySelection;
+
     // register your settings variables so the user can set them in the script settings
     //
     // unfortunately there is no QVariantHash in Qt, we only can use
@@ -54,6 +55,14 @@ Script {
             "description": "Please select the file:",
             "type": "file",
             "default": "pandoc",
+        },
+        {
+            "identifier": "mySelection",
+            "name": "I am an item selector",
+            "description": "Please select an item:",
+            "type": "selection",
+            "default": "option2",
+            "items": {"option1": "Text for option 1", "option2": "Text for option 2", "option3": "Text for option 3"},
         }
     ];
     
@@ -71,6 +80,7 @@ Script {
         script.log(myText);
         script.log(myInt);
         script.log(myFile);
+        script.log(mySelection);
     }
     
     /**