+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 524
+#define BUILD 525