+++ src/widgets/scriptsettingwidget.h
@@ -33,6 +33,8 @@ private slots:
 
     void on_booleanCheckBox_toggled(bool checked);
 
+    void on_selectionComboBox_currentIndexChanged(int index);
+
 private:
     Ui::ScriptSettingWidget *ui;
     QMap<QString, QVariant> _variableMap;